requirements.txt
slack-sdk==3.7.0
python-dotenv